# shared/specs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Edge AI Platform API
  version: 1.0.0
  description: |
    REST API for the Edge AI Platform web management interface.
    Provides device management, chat, and metrics access.
  contact:
    name: VLMChat Platform
  license:
    name: MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: http://localhost:3000
    description: Docker development server
security:
  - bearerAuth: []
...
paths:
...
  /api/devices:
    get:
      tags: [devices]
      summary: List all devices
      description: Get all edge devices registered to the current user
      operationId: listDevices
      parameters:
        - name: status
          in: query
          description: Filter devices by status
          required: false
          schema:
            type: string
            enum: [active, connected, all]
      responses:
        '200':
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    
    post:
      tags: [devices]
      summary: Register new device
      description: Create a new edge device registration
      operationId: createDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '200':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/devices/{deviceId}:
    get:
      tags: [devices]
      summary: Get device by ID
      description: Retrieve a specific device by its ID
      operationId: getDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Device details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          $ref: '#/components/responses/InternalError'
    
    delete:
      tags: [devices]
      summary: Delete device
      description: Remove a device registration
      operationId: deleteDevice
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Device deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      description: Device unique identifier
      schema:
        type: string
  schemas:
...
    Device:
      type: object
      description: |
        Device runtime state and identity.
        Lightweight response schema for device queries.
      required: [deviceId, name, status, createdAt]
      properties:
        deviceId:
          type: string
          description: Unique device identifier
        name:
          type: string
          description: Human-friendly device name
        status:
          type: string
          enum: [connected, disconnected, connecting, error]
          description: Current connection status
        userId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
          description: Owner user ID (null if unclaimed)
        lastSeen:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
          description: Last activity timestamp
        createdAt:
          type: string
          format: date-time
          description: Device creation timestamp
    DeviceCreate:
      description: |
        Admin device registration payload.
        Full hardware specification provided during manufacturing/provisioning.
      allOf:
        - $ref: '../schemas/register-v1.0.0.json'
...
  responses:
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: string
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
              message:
                type: string
                example: "Authentication required"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT session token from cookie-based authentication